Fix S3CrossAccountTrustRule granularity, rename s3_bucket_policy, add changelog to docs
